wip -- at this commit npm start fails but no tests pass... that's no bueno
